requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.7.0
joblib>=1.1
nltk>=3.7
pytest>=7.1
jupyter>=1.0.0
pandas>=1.5
torch==1.13.1
torchvision==0.14.1
transformers==4.26.1
openai
scikits-bootstrap
tqdm
